requirements.txt
accelerate
datasets
diffusers
google-genai>=1.0
openai
librosa
numpy
pandas
Pillow
safetensors
soundfile
torch
tqdm
transformers
yt-dlp
